api_wrappers_samples/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.16.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.4
Pillow==7.1.2
praw==7.0.0
prawcore==1.4.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
update-checker==0.17
uritemplate==3.0.1
urllib3==1.25.9
websocket-client==0.57.0
wordcloud==1.7.0
